+++ benchmarks/src/benchmark_reflect_serialization_twitter.hpp
@@ -1,0 +1,6 @@
+#ifndef BENCHMARK_REFLECT_SERIALIZATION_TWITTER_HPP
+#define BENCHMARK_REFLECT_SERIALIZATION_TWITTER_HPP
+
+#include "twitter_data.hpp"
+void bench_reflect_cpp(TwitterData& data);
+#endif
